move planning views to a separate file
