+++ timetable/university/urls.py
@@ -13,16 +13,19 @@ urlpatterns = patterns('',
     (r'^help/$', 'timetable.university.views.help'),
     (r'^about/$', 'timetable.university.views.about'),
     (r'^contacts/$', 'timetable.university.views.contacts'),
-    (r'^planning/(?P<term>\d+)/$', 'timetable.university.views.planning'),
-    (r'^planning/$', 'timetable.university.views.choose_term_for_planning'),
-    (r'^planning-light/$', 'timetable.university.views.planning_light'),
-    (r'^course-stats/$', 'timetable.university.views.course_stats'),
+    (r'^planning/(?P<term>\d+)/$',
+     'timetable.university.planning_views.planning'),
+    (r'^planning/$',
+     'timetable.university.planning_views.choose_term_for_planning'),
+    (r'^planning-light/$',
+     'timetable.university.planning_views.planning_light'),
+    (r'^course-stats/$', 'timetable.university.planning_views.course_stats'),
     (r'^planning-light/(?P<room_id>\d+)/$',
-     'timetable.university.views.planning_ajax'),
+     'timetable.university.planning_views.planning_ajax'),
     (r'^planning-light/room/(?P<room_id>\d+)/$',
-     'timetable.university.views.planning_light_room'),
+     'timetable.university.planning_views.planning_light_room'),
     ('^planning-light/room/(?P<room_id>\d+)/ajax/$',
-     'timetable.university.views.planning_room_ajax'),
+     'timetable.university.planning_views.planning_room_ajax'),
     (r'^choose-subjects/(?P<timetable_id>\d+)/$',
      'timetable.university.views.choose_subjects'),
     url(r'^render/(?P<encoded_groups>[\d/]+)/$',
